+++ src/test/java/org/lockss/spring/test/SpringLockssTestCase4.java
@@ -44,6 +44,7 @@ import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
 import org.lockss.app.LockssDaemon;
@@ -57,9 +58,11 @@ import org.lockss.util.TemplateUtil;
 import org.lockss.util.auth.*;
 import org.lockss.util.rest.RestUtil;
 import org.lockss.test.*;
-import org.skyscreamer.jsonassert.JSONAssert;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.core.env.ConfigurableEnvironment;
+import org.springframework.core.env.MapPropertySource;
+import org.springframework.core.env.MutablePropertySources;
 import org.springframework.http.*;
 import org.springframework.util.FileCopyUtils;
 import org.springframework.util.FileSystemUtils;
@@ -80,6 +83,9 @@ public abstract class SpringLockssTestCase4 extends LockssTestCase4 {
   public static final String PLATFORM_DISK_SPACE_CONFIG_FILENAME =
       "platform.txt";
 
+  public static final String MISC_CONFIG_FILENAME =
+      "misc.txt";
+
   /**
    * The name of the file with the UI port configuration template.
    */
@@ -94,7 +100,10 @@ public abstract class SpringLockssTestCase4 extends LockssTestCase4 {
   private static final Logger log = Logger.getLogger();
 
   @Autowired
-  private ApplicationContext appCtx;
+  private ConfigurableApplicationContext appCtx;
+
+  @Autowired(required = false)
+  private MockLockssDaemon mockLockssDaemon;
 
   // The path of a temporary directory where the test data will reside.
   private String tempDirPath = null;
@@ -102,6 +111,10 @@ public abstract class SpringLockssTestCase4 extends LockssTestCase4 {
   // The path to the configuration file with the platform disk space location
   // definition.
   private String platformDiskSpaceConfigPath = null;
+
+  // The path to the configuration file with miscellaneous configuration parameters
+  private String miscConfigPath;
+
 
   // The configuration file that specifies the UI port.
   private File uiPortConfigFile = null;
@@ -112,19 +125,42 @@ public abstract class SpringLockssTestCase4 extends LockssTestCase4 {
   // The configuration file that specifies the repository.
   private File repoConfigFile = null;
 
+  @Override
+  protected MockLockssDaemon newMockLockssDaemon() {
+    if (mockLockssDaemon == null) {
+      return super.newMockLockssDaemon();
+    }
+    log.info("newMockLockssDaemon() returning already created MockLockssDaemon");
+    return mockLockssDaemon;
+  }
+
   /** Create a fresh config manager.  This differs from what
    * LockssTestCase4 does in that it passes the ApplicationContext to
    * ConfigManager, and arranges for an initial config load to appear to
    * have been done. */
+  // Beware - most/all the Spring services tests start the app, which
+  // will create a new ConfigManager.  DO NOT use ConfigurationUtil to
+  // set any config params here, or in those tests' @Before methods,
+  // or at any point before the app is started.
   @Override
   protected ConfigManager makeConfigManager() {
-    ConfigManager mgr = ConfigManager.makeConfigManager(appCtx);
+    ConfigManager cfgMgr = ConfigManager.makeConfigManager(appCtx);
+
+    if (appCtx != null) {
+      ConfigurableEnvironment appCtxEnv = appCtx.getEnvironment();
+      MutablePropertySources propSrcs = appCtxEnv.getPropertySources();
+
+      Map<String, Object> myMap = new HashMap<>();
+      myMap.put("LockssConfigManager", cfgMgr);
+      propSrcs.addFirst(new MapPropertySource("MY_MAP", myMap));
+    }
+
     // Some Spring components, and esp. SpringAuthenticationFilter, wait
     // for the config to be loaded.  Which, in a test environment, doesn't
     // happen automatically.  Install an empty config so that all the
     // callbacks get called and waitConfig() returns.
     ConfigurationUtil.installConfig(ConfigManager.newConfiguration());
-    return mgr;
+    return cfgMgr;
   }
 
   /**
@@ -310,6 +346,33 @@ public abstract class SpringLockssTestCase4 extends LockssTestCase4 {
     return platformDiskSpaceConfigPath;
   }
 
+
+  public String getMiscConfigPath() {
+    return miscConfigPath;
+  }
+
+  public void addParamToMiscConfig(String key, String value)
+      throws IOException {
+
+    if (miscConfigPath == null) {
+      // Get the path of the misc directory
+      String miscDirPath = getTempDir("misc").getAbsolutePath();
+      if (log.isDebug3()) log.debug3("miscDirPath = " + miscDirPath);
+
+      // The path to the file.
+      miscConfigPath = miscDirPath + File.separator + MISC_CONFIG_FILENAME;
+    }
+
+    String param = key + "=" + value + "\n";
+
+    // Append to the misc file.
+    Files.write(Paths.get(miscConfigPath),
+        param.getBytes(), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+
+    if (log.isDebug2()) log.debug2("miscConfigPath = "
+        + miscConfigPath);
+  }
+
   /**
    * Copies a file or directory to the temporary directory.
    *
@@ -379,6 +442,7 @@ public abstract class SpringLockssTestCase4 extends LockssTestCase4 {
   protected File getUiPortConfigFile() {
     return uiPortConfigFile;
   }
+
 
   /**
    * Creates the configuration file that specifies the database properties.
@@ -504,7 +568,7 @@ public abstract class SpringLockssTestCase4 extends LockssTestCase4 {
   }
 
   /** Intended for tests of code that normally runs in on-demand AU
-   * creation mode (e.g., mdq & mdx services), to make the tests work in
+   * creation mode (e.g., md service), to make the tests work in
    * startAllAus mode.  Creates an AU from config inferred from the AUID,
    * iff the AU doesn\"t already exist and the daemon is not running in
    * on-demand mode. */